backend/analytics-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
python-dotenv==1.1.0
PyJWT==2.10.1